Refactor more code, modularize models
