+++ routes/user.js
@@ -4,7 +4,7 @@ const express = require('express');
 const router = express.Router();
 const morgan = require('morgan');
 const mongoose = require('mongoose');
-const {User} = require('../models');
+const {User} = require('../models/user-model');
 const bodyParser = require('body-parser');
 
 mongoose.Promise = global.Promise;
