+++ seeddata.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { User, Raid } = require('./models');
+const { User, Raid } = require('./model/models');
 const mongoose = require('mongoose');
 const {DATABASE_URL} = require('./config');
 const faker = require('faker');
